Revert "UP TO 20"
nfsw Access Violation in DEP was traced to testing with 3c85dda reverted which worked.
Initial test build reverted all Affinity changes to validate the ptoential culprit change, and was then walked forward to include only the original Affinity update change. The "Up to 20 processors" part continued to break launching the game on a Ryzen 9 3900X.