Move Unused Code to Gist and Update legacy_modnet URL
This update is meant to clean up MainScreen.cs since there is quite a bit of old code that is unused and messy when trying to debug. Instead of fully removing it, I went ahead and moved it to Gist with a link of the unused code.

Also went ahead and updated "legacy_modnet" URL from "cdn.mtntr.pl" to "cdn.worldunited.gg" this might explain some issue with users disabling proxy.