Add "LauncherPath" to UI
Adds a "Launcher Folder:" label and a clickable explorer "LauncherPathCurrent"; so users that may have misplaced where they put the launcher can acces the folder from within the launcher (in the event the propertiees of a shortcut escapes them).

OCD may have made me reposition a lot of things to align with their viusal location order, and in files maybe not necessarily meant to be human readable, but I digress.