Update VerifyHash.cs
Huge thanks to Metonator.
Validator now works for establishing files as requested in the checksums.dat, recording and recorgnizing any "missing" files as well as corrupted ones.

ToDo:
Tie this into the UI with a progressbar indication and .ToUpper feedback on what/where it's scanning, as well as what it has to download/download proogress.
Logging: Not only fix state/status logging, but if possible (so as not to bleat the launcher.log file) log Validation.log explicitly for the output.