Fix Firewall Service and Profile Checks. Also Removed Splash Screen
Fixed:
Firewall rules running even when the Service and/or Profile is not active. Added a second check to prevent these crashes and added additional logging to display the firewall status

Changed:
Moved the check where it checks the launcher or game is running and prevents the launcher to open another instance.

Removed:
Splash Screen has been nuked. Since it adds an additional amount of time on launch even of fast SSDs so its temporarily been removed until a better solution has been found.