Updated Proxy, Moved WebTimeout and Other
Reverted:
The current Compress Response might be posing issues so, lets revert to an older version that Allocates Memory. This time I made notes to why that system was used instead of the default method.

Moved:
SecurityProtocol and Expect100Continue to Program.cs since they are only needed to be called once. In Addition, I've changed the `==` to `|=` since `|=`adds on to the SecurityProtocol that is support on the Framework instead of Overriding it. Hopefully this change should not cause changes.

WebRequest was Moved just to move it.  (:weirdchamp:)