Revert a136b1a
This was causing the Dropdown List to remain static until a server was selected.

Given that the Status indication this is trying to solve is at best an edge case where a Custom Server shares "id" with an API List server (Linking the Status Polling between them when either is selected), I'm going to chalk this one to a "Won't Fix" for now.