GSI Update
We shouldn't have to clean up broken GSI data.

modernAuthSupport is a bool, not a string, so we shouldn't need to stringify/handle it as one it (just to re-/bool it)

modernAuthSupport also has nothing to do with whether or not Proxy is forced on for https connections.

I made sure that proxy was turned off for http if the enforceServerProxy wasn't present, even if a the launcher started loaded to an https (or had proxy Disabled checked on an https server selected) and then switched to an http to login.

It should still enforceLauncherProxy on an http server though if it's set in GSI (and as a bool, because if they put it up wrong that is on them)

This also means the Launcher GSI can treat it as `bool` properly instead of as `bool?`