Change Servers-Custom and restore Settings
Settings.ini should be fine for storing with the Launcher, esp for people that might do a Launcher+GameFiles per server, since nothing account specific is detailed in it.

Servers-Custom however -could- be customized to each player, especially with Account.ini storing the last connected server details; on a multi-user PC this may not be desirable.
(Made sure it was Linux Friendly, I hope)

Additionally, AddServer.cs shouldn't have to invoke checking/moving or creating the JSON as it's already been executed by Program.cs, which will either move or create or bail well before we get to ever displaying ServerSelect (1st Run) or the player interacting with MainScreen.